Fix Windows path issue for MenuExample imports
